consumer/kafka_consumer.py
from kafka import KafkaConsumer
from json import loads

consumer = KafkaConsumer(
    'blocks',
    bootstrap_servers=['127.0.0.1:9092'],
    auto_offset_reset='earliest',
    enable_auto_commit=True,
    group_id='my-group-2',
    value_deserializer=lambda x: loads(x.decode('utf-8')))

for message in consumer:
    message = message.value
    print(message)
