requirements.txt
kafka-python==2.0.2
numpy==1.23.4
pandas==1.5.1
py4j==0.10.9.5
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
